+++ src/theme/mod.rs
@@ -1,2 +1,3 @@
+pub mod installer;
 pub mod loader;
 pub mod models;